+++ h/meeting_service_components/meeting_configuration_interface.h
@@ -1,7 +1,6 @@
 /*!
 * \file meeting_configuration_interface.h
 * \brief Configuration of Meeting Service Interface
-* 
 */
 #ifndef _MEETING_Configuration_INTERFACE_H_
 #define _MEETING_Configuration_INTERFACE_H_
@@ -19,20 +18,106 @@ public:
 	enum RequiredInfoType
 	{
 		REQUIRED_INFO_TYPE_NONE,
-		REQUIRED_INFO_TYPE_Password, ///< if you want to join meeting, you need input password
-		REQUIRED_INFO_TYPE_Password4WrongPassword,///< you input a wrong password, please input again
-		REQUIRED_INFO_TYPE_PasswordAndScreenName,///< if you want to join meeting, you need input password and screen name
+		REQUIRED_INFO_TYPE_Password, ///< if you want to join meeting, you need input password with api InputMeetingPasswordAndScreenName
+		REQUIRED_INFO_TYPE_Password4WrongPassword,///< you input a wrong password, please input again with api InputMeetingPasswordAndScreenName
+		REQUIRED_INFO_TYPE_PasswordAndScreenName,///< if you want to join meeting, you need input password and screen name with api InputMeetingPasswordAndScreenName
+		REQUIRED_INFO_TYPE_ScreenName,///< if you want to join meeting, you need input screen name with api InputMeetingScreenName
 	};
 
 	/// \brief get the required information you need to input.
 	virtual RequiredInfoType GetRequiredInfoType() = 0;
 
-	/// \brief Input the meeting password and screen name.
+	/// \brief Input the meeting password and screen name, and release this object
 	virtual bool InputMeetingPasswordAndScreenName(const wchar_t* meetingPassword, const wchar_t* screenName) = 0;
 
-	// \brief cancel to input the meeting password and screen name.
+	/// \brief Input the meeting screen name, and release this object
+	virtual bool InputMeetingScreenName(const wchar_t* screenName) = 0;
+
+	// \brief cancel to input the meeting password and screen name, and release this object
 	virtual void Cancel() = 0;
 	virtual ~IMeetingPasswordAndScreenNameHandler() {};
+};
+
+/// \brief Register webinar Handler
+///
+class IWebinarNeedRegisterHandler
+{
+public:
+	enum WebinarNeedRegisterType
+	{
+		WebinarReg_NONE,
+		WebinarReg_By_Register_Url, //dynamic_cast to IWebinarNeedRegisterHandlerByUrl
+		WebinarReg_By_Email_and_DisplayName,//dynamic_cast to IWebinarNeedRegisterHandlerByEmail
+	};
+
+	virtual WebinarNeedRegisterType GetWebinarNeedRegisterType() = 0;
+	virtual ~IWebinarNeedRegisterHandler() {};
+};
+
+/// \brief Register webinar with url Handler
+///
+class IWebinarNeedRegisterHandlerByUrl : public IWebinarNeedRegisterHandler
+{
+public:
+	/// \brief Get the webinar register url
+	virtual const wchar_t* GetWebinarRegisterUrl() = 0;
+
+	/// \brief Release object
+	virtual void Release() = 0;
+	virtual ~IWebinarNeedRegisterHandlerByUrl() {};
+};
+
+/// \brief Register webinar with email and screen name Handler
+///
+class IWebinarNeedRegisterHandlerByEmail : public IWebinarNeedRegisterHandler
+{
+public:
+	/// \brief Input the register email and screen name, and release this object
+	virtual SDKError InputWebinarRegisterEmailAndScreenName(const wchar_t* email, const wchar_t* screenName) = 0;
+
+	// \brief cancel to input the the register email and screen name, and release this object
+	virtual void Cancel() = 0;
+	virtual ~IWebinarNeedRegisterHandlerByEmail() {};
+};
+
+/// \brief End other meeting to join the new meeting Handler
+///
+class IEndOtherMeetingToJoinMeetingHandler
+{
+public:
+	virtual bool EndOtherMeeting() = 0;
+	virtual void Cancel() = 0;
+	virtual ~IEndOtherMeetingToJoinMeetingHandler() {};
+};
+
+
+/// \brief Free meeting ending reminder Handler
+///
+class IFreeMeetingEndingReminderHandler
+{
+public:
+	enum FreeMeetingEndingReminderType
+	{
+		FreeMeetingEndingReminder_NONE,
+		FreeMeetingEndingReminder_Can_UpgradeMeeting_ToPro_Once,
+		FreeMeetingEndingReminder_Can_UpgradeAccount,
+		FreeMeetingEndingReminder_Reminder,
+		FreeMeetingEndingReminder_UpgradeMeeting_ToPro_Failed,
+		FreeMeetingEndingReminder_UpgradeMeeting_ToPro_Success,
+	};
+
+	/// \brief get the reminder type.
+	virtual FreeMeetingEndingReminderType GetType() = 0;
+
+	/// \brief Upgrade meeting and release this object, only for FreeMeetingEndingReminder_Can_UpgradeMeeting_ToPro_Once
+	virtual SDKError UpgradeMeeting() = 0;
+
+	/// \brief Upgrade account and release this object, only for FreeMeetingEndingReminder_Can_UpgradeAccount
+	virtual SDKError UpgradeAccount() = 0;
+
+	// \brief cancel, and release this object
+	virtual SDKError Cancel() = 0;
+	virtual ~IFreeMeetingEndingReminderHandler() {};
 };
 
 /// \brief Meeting Configuration Event callback
@@ -48,6 +133,18 @@ public:
 	/// \param bShow Specifies show or hide your own airplay instruction window.
 	/// \param airhostName the airhost name displayed on ios device
 	virtual void onAirPlayInstructionWndNotification(bool bShow, const wchar_t* airhostName) = 0;
+
+	/// \brief webinar need register callback
+	/// \param handler_ which will be invoked for when this message type is received.
+	virtual void onWebinarNeedRegisterNotification(IWebinarNeedRegisterHandler* handler_) = 0;
+
+	/// \brief meeting need to end other meeting to join this meeting callback
+	/// \param handler_ which will be invoked for when this message type is received.
+	virtual void onEndOtherMeetingToJoinMeetingNotification(IEndOtherMeetingToJoinMeetingHandler* handler_) = 0;
+
+	/// \brief the notification of free meeting will ending callback
+	/// \param handler_ which will be invoked for when this message type is received.
+	virtual void onFreeMeetingEndingReminderNotification(IFreeMeetingEndingReminderHandler* handler_) = 0;
 };
 
 /// \brief Meeting Configuration Interface
@@ -135,6 +232,21 @@ public:
 	/// if disable this option, you need to handle onInputMeetingPasswordAndScreenNameNotification event callback.
 	virtual void EnableInputMeetingPasswordDlg(bool bEnable) = 0;
 
+	/// \brief config hide or show input meeting screen name dialog
+	/// \param bEnable specify hide or show input meeting screen name dialog, default is enable.
+	/// if disable this option, you need to handle onInputMeetingPasswordAndScreenNameNotification event callback.
+	virtual void EnableInputMeetingScreenNameDlg(bool bEnable) = 0;
+
+	/// \brief Redirect webinar register logic to IMeetingConfigurationEvent
+	/// \param bRedirect Redirect or not. if true, you will recv onWebinarNeedRegisterNotification in IMeetingConfigurationEvent
+	/// \default is disable.
+	virtual void RedirectWebinarNeedRegister(bool bRedirect) = 0;
+
+	/// \brief Redirect end other to IMeetingConfigurationEvent
+	/// \param bRedirect Redirect or not. if true, you will recv onEndOtherMeetingToJoinMeetingNotification in IMeetingConfigurationEvent
+	/// \default is disable.
+	virtual void RedirectEndOtherMeeting(bool bRedirect) = 0;
+
 	// \brief Enable enter and exit full screen button on meeting ui. 
 	/// \param bEnable Specifies enter and exit full screen button show or not in meeting window, default is enable.
 	virtual void EnableEnterAndExitFullScreenButtonOnMeetingUI(bool bEnable) = 0;
@@ -202,6 +314,47 @@ public:
 	/// \brief Enable the length limitation of meeting number
 	/// \param bEnable specify this feature enable or not, default is enable.
 	virtual void EnableLengthLimitationOfMeetingNumber(bool bEnable) = 0;
+
+	/// \brief Enable IOS device share
+	/// \param bEnable specify this feature enable or not, default is enable.
+	virtual void EnableShareIOSDevice(bool bEnable) = 0;
+
+	/// \brief Enable White Board share
+	/// \param bEnable specify this feature enable or not, default is enable.
+	virtual void EnableShareWhiteBoard(bool bEnable) = 0;
+
+	/// \brief Always show app icon on taskbar list
+	/// \param bAlwaysShow specify always show or not, default is not.
+	virtual void AlwaysShowIconOnTaskBar(bool bAlwaysShow) = 0;
+
+	/// \brief Redirect free meeting will end message to IMeetingConfigurationEvent
+	/// \param bRedirect Redirect or not. if true, you will recv onFreeMeetingEndingReminderNotification in IMeetingConfigurationEvent
+	virtual void RedirectFreeMeetingEndingReminderDlg(bool bRedirect) = 0;
+
+	/// \brief Show the 'Share computer sound' checkbox
+	/// \param bShow specify to show the checkbox or not, default is true.
+	virtual void SetShowAudioUseComputerSoundChkbox(bool bShow) = 0;
+
+	/// \brief Show the 'Optimize for full screen video clip' checkbox
+	/// \param bShow specify to show the checkbox or not, default is true.
+	virtual void SetShowVideoOptimizeChkbox(bool bShow) = 0;
+	
+	/// \brief Show the 'Phone Call' tab
+	/// \param bShow specify to hide the tab when it's value is false, default is true.
+	virtual void SetShowCallInTab(bool bShow) = 0;
+
+	/// \brief Show the 'Call Me' tab
+	/// \param bShow specify to hide the tab when it's value is false, default is true.
+	virtual void SetShowCallMeTab(bool bShow) = 0;
+
+	/// \brief Force disable multiple share
+	/// \param bDisable specify this feature disable or not, default is enable.
+	virtual void ForceDisableMultiShare(bool bDisable) = 0;
+
+	/// \brief Always show the Meeting ID information on the main video-window's title bar 
+	/// \param bAlwaysShow specify to show the meeting id information on the windows's title bar always or not. If bAlwaysShow if false, the behavior will follow Zoom's default logic.Default is false.
+	virtual void SetAlwaysShowMeetingIDOnTitle(bool bAlwaysShow) = 0;
 };
+
 END_ZOOM_SDK_NAMESPACE
 #endif