+++ h/meeting_service_components/meeting_sharing_interface.h
@@ -126,6 +126,11 @@ public:
 	///If the function fails, the return value is not SDKErr_Success. To get extended error information, refer to SDKError enum.
 	virtual SDKError StartAirPlayShare() = 0;
 
+	/// \brief Start share with White board.
+	/// \return If the function succeeds, the return value is SDKErr_Success.
+	///If the function fails, the return value is not SDKErr_Success. To get extended error information, refer to SDKError enum.
+	virtual SDKError StartWhiteBoardShare() = 0;
+
 	/// \brief Show sharing app select window.
 	/// \return If the function succeeds, the return value is SDKErr_Success.
 	///If the function fails, the return value is not SDKErr_Success. To get extended error information, refer to SDKError enum.