+++ h/meeting_service_components/meeting_waiting_room_interface.h
@@ -1,7 +1,6 @@
 /*!
 * \file meeting_waiting_room_interface.h
 * \brief WaitingRoom of Meeting Service Interface
-* 
 */
 #ifndef _MEETING_WaitingRoom_INTERFACE_H_
 #define _MEETING_WaitingRoom_INTERFACE_H_